Added react router, added bg video and Header
React router was used to create the links for the additional pages
although they do not exist yet. A Header component was made to add the
links to and they can be clicked and change the URL but not the page
yet.

The bg video should be stored in /src/assets and named placeholder.mp4

The Playfair font family was imported and can be used in Tailwind.